Report errors/warnings on fallback to copy in source package install on Windows. Do not copy when rename removed the origin or created the target.
git-svn-id: https://svn.r-project.org/R/trunk@76888 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/tools/R/install.R
@@ -1673,10 +1673,15 @@ if(FALSE) {
             if (WINDOWS) {
                 unlink(final_instdir, recursive = TRUE) # needed for file.rename
                 if (!file.rename(instdir, final_instdir)) {
-                    message("WARNING: moving package to final location failed, copying instead")
-                    file.copy(instdir, dirname(final_instdir), recursive = TRUE,
-                              copy.date = TRUE)
-                    unlink(instdir, recursive = TRUE)
+                    if (dir.exists(instdir) && !dir.exists(final_instdir)) {
+                        message("WARNING: moving package to final location failed, copying instead")
+                        ret <- file.copy(instdir, dirname(final_instdir),
+                                         recursive = TRUE, copy.date = TRUE)
+                        if (any(!ret))
+                            errmsg("   copying to final location failed")
+                        unlink(instdir, recursive = TRUE)
+                    } else
+                        errmsg("   moving to final location failed")
                 }
             } else {
                 patch_rpaths()
